% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_gadget.r
\name{new_plot_gadget}
\alias{new_plot_gadget}
\title{Title}
\usage{
new_plot_gadget(data)
}
\arguments{
\item{data}{A data frame to visualize.}
}
\value{
A shiny gadget interface.
}
\description{
Title
}
\examples{
\dontrun{
  # Launch the interactive gadget
  new_plot_gadget(mtcars)
}
}
